Moved data functions to data-utils to facilitate easier testing in the future. For now, it seems like I cannot properly mock up a FileReader / FileList.
