+++ src/context/index.ts
@@ -1,4 +1,3 @@
 export { DataProvider } from "./provider";
 export { useData, useDataDispatch } from "./context";
-export type { State } from "./types";
-export type { User } from "./types.ts";
+export type { State, Action, User } from "./types";
