+++ src/routes/root/header.tsx
@@ -1,4 +1,4 @@
-import { DataFunctions } from "./data-functions";
+import DataFunctions from "./data-functions";
 
 export const Header = () => (
   <>
